+++ js/form.js
@@ -8,9 +8,7 @@
 //     эффект сбрасывается на «Оригинал»;
 //     поля для ввода хэш-тегов и комментария очищаются;
 //     поле загрузки фотографии, стилизованное под букву «О» в логотипе, очищается.
-const VALID_SYMBOLS = /^#[a-zа-яё0-9]{1,19}$/i;
-const MAX_HASHTAG_COUNT = 5;
-const TAG_ERROR_TEXT = 'Некорректно записаны хэштеги';
+
 const form = document.querySelector('.img-upload__form');
 const overlay = document.querySelector('.img-upload__overlay');
 const body = document.querySelector('.body');
@@ -19,11 +17,6 @@ const fileField = document.querySelector('#upload-file');
 const hashtagField = document.querySelector('.text__hashtags');
 const commentField = document.querySelector('.text__descriptions');
 
-const pristine = new Pristine(form, {
-  classTo: 'img-upload__field-wrapper',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextClass: 'img-upload__field-wrapper',
-});
 
 const showModal = () =>{
   overlay.classList.remove('.hidden');
@@ -33,7 +26,6 @@ const showModal = () =>{
 
 const hideModal = () =>{
   form.reset();
-  pristine.reset();
   overlay.classList.add('hidden');
   body.classList.remove('modal-open');
   document.removeEventListener('keydown', onDocumentKeydown);
@@ -57,34 +49,12 @@ const onFileInputChange = () =>{
   showModal();
 };
 
-const isValidTag = (tag) => VALID_SYMBOLS.test(tag);
-
-const hasValidCount = (tags) => tags.length <= MAX_HASHTAG_COUNT;
-
-const hasUniqueTags = (tags) =>{
-  const lowerCaseTags = tags.map((tag) => tag.toLowerCase());
-  return lowerCaseTags.length === new Set(lowerCaseTags).size;
-};
-
-const validateTags = (value) =>{
-  const tags = value
-    .trim()
-    .split(' ')
-    .filter((tag) => tag.trim().length);
-  return hasValidCount(tags) && hasUniqueTags(tags) && tags.every(isValidTag);
-};
-
-pristine.addValidator(
-  hashtagField,
-  validateTags,
-  TAG_ERROR_TEXT
-);
-
 const onFormSubmit = (evt) =>{
   evt.preventDefault();
-  pristine.validate();
 };
 
 fileField.addEventListener('change', onFileInputChange);
 cancelButton.addEventListener('click',onCancelButtonClick);
 form.addEventListener('submit', onFormSubmit);
+
+export {form,showModal, hideModal};
